fail when @setting is defined in a non-extension class
