+++ plugins/autodoc/autodoc-processor/src/test/java/org/eclipse/edc/plugins/autodoc/core/processor/test/NotExtensionWithSetting.java
@@ -12,11 +12,11 @@
  *
  */
 
-package org.eclipse.edc.plugins.autodoc.core.processor.testextensions;
+package org.eclipse.edc.plugins.autodoc.core.processor.test;
 
-import org.eclipse.edc.runtime.metamodel.annotation.Inject;
+import org.eclipse.edc.runtime.metamodel.annotation.Setting;
 
-public class NotAnExtension {
-    @Inject(required = false)
-    private OptionalService optionalService;
+public class NotExtensionWithSetting {
+    @Setting("the setting must stay in a ServiceExtension class")
+    private static final String UNEXPECTED_SETTING = "any";
 }
